requirements.txt
Flask
python-dotenv

langchain
langchain-core
langchain-google-genai
langchain-text-splitters
langchain-community
langchain-pinecone
pinecone-client
bs4
beautifulsoup4


